Move GsfElectronAlgoHeavyObjectCache together with GsfElectronAlgo
