+++ RecoEgamma/EgammaElectronAlgos/src/ElectronClassification.cc
@@ -2,8 +2,6 @@
 
 #include "DataFormats/EgammaReco/interface/SuperCluster.h"
 
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-
 //===================================================================
 // Author: Federico Ferri - INFN Milano, Bicocca university
 // 12/2005
